+++ index.html
@@ -231,6 +231,9 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
+  <script src="main.js" type="module"></script>
+  <script src="data.js"></script>
+  <script src="util.js" type="module"></script>
 
 </body>
 </html>
