+++ package.json
@@ -24,5 +24,11 @@
   "engines": {
     "node": "16.x",
     "npm": "8.x"
-  }
+   },
+
+
+  "type": "module"
+
+
 }
+
